+++ src/Resources/contao/languages/de/tl_adressen.php
@@ -3,164 +3,98 @@
 /**
  * Backend-Modul: Übersetzungen im Eingabeformular
  */
-$GLOBALS["TL_LANG"]["tl_adressen"]["id"] = array("ID", "ID dieser Adresse");
-$GLOBALS["TL_LANG"]["tl_adressen"]["tstamp"] = array("Letzte Änderung", "Zeitstempel der letzten Änderung");
+$GLOBALS['TL_LANG']['tl_adressen']['id'] = array('ID', 'ID dieser Adresse');
+$GLOBALS['TL_LANG']['tl_adressen']['tstamp'] = array('Letzte Änderung', 'Zeitstempel der letzten Änderung');
 
-$GLOBALS["TL_LANG"]["tl_adressen"]["person_legende"] = "Name und Firma";
-$GLOBALS["TL_LANG"]["tl_adressen"]["nachname"] = array("Nachname", "Geben Sie hier den Nachnamen ein. Alternativ wäre auch der Vereins- oder Firmenname möglich.");
-$GLOBALS["TL_LANG"]["tl_adressen"]["vorname"] = array("Vorname", "Geben Sie hier den Vornamen ein.");
-$GLOBALS["TL_LANG"]["tl_adressen"]["titel"] = array("Titel", "Geben Sie hier den Titel ein, z.B. Dr. oder Prof.Dr.");
-$GLOBALS["TL_LANG"]["tl_adressen"]["firma"] = array("Firma", "Geben Sie hier den Firmennamen ein.");
-$GLOBALS["TL_LANG"]["tl_adressen"]["club"] = array("Verein", "Geben Sie hier den Verein der Person ein.");
+$GLOBALS['TL_LANG']['tl_adressen']['person_legende'] = 'Name und Firma';
+$GLOBALS['TL_LANG']['tl_adressen']['nachname'] = array('Nachname', 'Geben Sie hier den Nachnamen ein. Alternativ wäre auch der Vereins- oder Firmenname möglich.');
+$GLOBALS['TL_LANG']['tl_adressen']['vorname'] = array('Vorname', 'Geben Sie hier den Vornamen ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['titel'] = array('Titel', 'Geben Sie hier den Titel ein, z.B. Dr. oder Prof.Dr.');
+$GLOBALS['TL_LANG']['tl_adressen']['firma'] = array('Firma', 'Geben Sie hier den Firmennamen ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['club'] = array('Verein', 'Geben Sie hier den Verein der Person ein.');
 
-$GLOBALS["TL_LANG"]["tl_adressen"]["adresse_legende"] = "Anschrift";
-$GLOBALS["TL_LANG"]["tl_adressen"]["strasse_view"] = array("Straße veröffentlichen", "Straße im Frontend anzeigen. Wird nur angezeigt, wenn auch PLZ und Ort öffentlich sind.");
-$GLOBALS["TL_LANG"]["tl_adressen"]["strasse"] = array("Straße", "Geben Sie hier die Straße ein.");
-$GLOBALS["TL_LANG"]["tl_adressen"]["ort_view"] = array("PLZ und Ort veröffentlichen", "PLZ und Ort im Frontend anzeigen");
-$GLOBALS["TL_LANG"]["tl_adressen"]["plz"] = array("PLZ", "Geben Sie hier die Postleitzahl ein.");
-$GLOBALS["TL_LANG"]["tl_adressen"]["ort"] = array("Ort", "Geben Sie hier den Ort ein.");
+$GLOBALS['TL_LANG']['tl_adressen']['adresse_legende'] = 'Anschrift';
+$GLOBALS['TL_LANG']['tl_adressen']['strasse_view'] = array('Straße veröffentlichen', 'Straße im Frontend anzeigen. Wird nur angezeigt, wenn auch PLZ und Ort öffentlich sind.');
+$GLOBALS['TL_LANG']['tl_adressen']['strasse'] = array('Straße', 'Geben Sie hier die Straße ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['ort_view'] = array('PLZ und Ort veröffentlichen', 'PLZ und Ort im Frontend anzeigen');
+$GLOBALS['TL_LANG']['tl_adressen']['plz'] = array('PLZ', 'Geben Sie hier die Postleitzahl ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['ort'] = array('Ort', 'Geben Sie hier den Ort ein.');
 
-$GLOBALS["TL_LANG"]["tl_adressen"]["adressen_legend"] = "Anschriften (neues Format)";
-$GLOBALS["TL_LANG"]["tl_adressen"]["adressen"] = array("Adressen", "");
-$GLOBALS["TL_LANG"]["tl_adressen"]["adressen_public_plzort"] = array("PLZ/Ort öffentlich", "PLZ und Ort im Frontend anzeigen");
-$GLOBALS["TL_LANG"]["tl_adressen"]["adressen_plz"] = array("PLZ", "Geben Sie hier die Postleitzahl ein.");
-$GLOBALS["TL_LANG"]["tl_adressen"]["adressen_ort"] = array("Ort", "Geben Sie hier den Ort ein.");
-$GLOBALS["TL_LANG"]["tl_adressen"]["adressen_public_str"] = array("Straße öffentlich", "Straße im Frontend anzeigen. Wird nur angezeigt, wenn auch PLZ und Ort öffentlich sind.");
-$GLOBALS["TL_LANG"]["tl_adressen"]["adressen_strasse"] = array("Straße", "Geben Sie hier die Straße ein.");
-$GLOBALS["TL_LANG"]["tl_adressen"]["adressen_googlemap"] = array("Googlemap verlinken", "Adresse mit GoogleMap verlinken");
+$GLOBALS['TL_LANG']['tl_adressen']['adressen_legend'] = 'Anschriften (neues Format)';
+$GLOBALS['TL_LANG']['tl_adressen']['adressen'] = array('Adressen', '');
+$GLOBALS['TL_LANG']['tl_adressen']['adressen_public_plzort'] = array('PLZ/Ort öffentlich', 'PLZ und Ort im Frontend anzeigen');
+$GLOBALS['TL_LANG']['tl_adressen']['adressen_plz'] = array('PLZ', 'Geben Sie hier die Postleitzahl ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['adressen_ort'] = array('Ort', 'Geben Sie hier den Ort ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['adressen_public_str'] = array('Straße öffentlich', 'Straße im Frontend anzeigen. Wird nur angezeigt, wenn auch PLZ und Ort öffentlich sind.');
+$GLOBALS['TL_LANG']['tl_adressen']['adressen_strasse'] = array('Straße', 'Geben Sie hier die Straße ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['adressen_googlemap'] = array('Googlemap verlinken', 'Adresse mit GoogleMap verlinken');
 
-$GLOBALS["TL_LANG"]["tl_adressen"]["funktionen_legende"] = "Funktionen und Kategorien";
-$GLOBALS["TL_LANG"]["tl_adressen"]["wertungsreferent"] = array("Wertungsreferent", "Wählen Sie hier die Wertungsreferate der Person aus.");
-$GLOBALS["TL_LANG"]["tl_adressen"]["funktionen"] = array("Kategorien", "Wählen Sie hier die Kategorien für die Person aus.");
+$GLOBALS['TL_LANG']['tl_adressen']['funktionen_legende'] = 'Funktionen und Kategorien';
+$GLOBALS['TL_LANG']['tl_adressen']['wertungsreferent'] = array('Wertungsreferent', 'Wählen Sie hier die Wertungsreferate der Person aus.');
+$GLOBALS['TL_LANG']['tl_adressen']['funktionen'] = array('Kategorien', 'Wählen Sie hier die Kategorien für die Person aus.');
 
-$GLOBALS["TL_LANG"]["tl_adressen"]["telefon_legende"] = "Telefon";
-$GLOBALS["TL_LANG"]["tl_adressen"]["telefon_view"] = array("Telefon veröffentlichen", "Alle Telefonnummern im Frontend anzeigen");
-$GLOBALS["TL_LANG"]["tl_adressen"]["telefon1"] = array("Telefon 1", "Geben Sie hier die erste Telefonnummer ein.");
-$GLOBALS["TL_LANG"]["tl_adressen"]["telefon2"] = array("Telefon 2", "Geben Sie hier die zweite Telefonnummer ein.");
-$GLOBALS["TL_LANG"]["tl_adressen"]["telefon3"] = array("Telefon 3", "Geben Sie hier die dritte Telefonnummer ein.");
-$GLOBALS["TL_LANG"]["tl_adressen"]["telefon4"] = array("Telefon 4", "Geben Sie hier die vierte Telefonnummer ein.");
+$GLOBALS['TL_LANG']['tl_adressen']['telefon_legende'] = 'Telefon';
+$GLOBALS['TL_LANG']['tl_adressen']['telefon_view'] = array('Telefon veröffentlichen', 'Alle Telefonnummern im Frontend anzeigen');
+$GLOBALS['TL_LANG']['tl_adressen']['telefon1'] = array('Telefon 1', 'Geben Sie hier die erste Telefonnummer ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['telefon2'] = array('Telefon 2', 'Geben Sie hier die zweite Telefonnummer ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['telefon3'] = array('Telefon 3', 'Geben Sie hier die dritte Telefonnummer ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['telefon4'] = array('Telefon 4', 'Geben Sie hier die vierte Telefonnummer ein.');
 
-$GLOBALS["TL_LANG"]["tl_adressen"]["telefone_legend"] = "Telefonnummern (neues Format)";
-$GLOBALS["TL_LANG"]["tl_adressen"]["telefone"] = array("Nummern", "");
-$GLOBALS["TL_LANG"]["tl_adressen"]["telefone_public_num"] = array("Nummer öffentlich", "Telefonnummer im Frontend anzeigen");
-$GLOBALS["TL_LANG"]["tl_adressen"]["telefone_nummer"] = array("Nummer", "Geben Sie hier die Telefonnummer ein.");
+$GLOBALS['TL_LANG']['tl_adressen']['telefone_legend'] = 'Telefonnummern (neues Format)';
+$GLOBALS['TL_LANG']['tl_adressen']['telefone'] = array('Nummern', '');
+$GLOBALS['TL_LANG']['tl_adressen']['telefone_public_num'] = array('Nummer öffentlich', 'Telefonnummer im Frontend anzeigen');
+$GLOBALS['TL_LANG']['tl_adressen']['telefone_nummer'] = array('Nummer', 'Geben Sie hier die Telefonnummer ein.');
 
-$GLOBALS["TL_LANG"]["tl_adressen"]["telefax_legende"] = "Telefax";
-$GLOBALS["TL_LANG"]["tl_adressen"]["telefax_view"] = array("Telefax veröffentlichen", "Telefaxnummern im Frontend anzeigen");
-$GLOBALS["TL_LANG"]["tl_adressen"]["telefax1"] = array("Telefax 1", "Geben Sie hier die erste Telefaxnummer ein.");
-$GLOBALS["TL_LANG"]["tl_adressen"]["telefax2"] = array("Telefax 2", "Geben Sie hier eine zweite Telefaxnummer ein.");
+$GLOBALS['TL_LANG']['tl_adressen']['telefax_legende'] = 'Telefax';
+$GLOBALS['TL_LANG']['tl_adressen']['telefax_view'] = array('Telefax veröffentlichen', 'Telefaxnummern im Frontend anzeigen');
+$GLOBALS['TL_LANG']['tl_adressen']['telefax1'] = array('Telefax 1', 'Geben Sie hier die erste Telefaxnummer ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['telefax2'] = array('Telefax 2', 'Geben Sie hier eine zweite Telefaxnummer ein.');
 
-$GLOBALS["TL_LANG"]["tl_adressen"]["email_legende"] = "E-Mail";
-$GLOBALS["TL_LANG"]["tl_adressen"]["email_view"] = array("E-Mail veröffentlichen", "E-Mail-Adressen im Frontend anzeigen");
-$GLOBALS["TL_LANG"]["tl_adressen"]["email1"] = array("E-Mail 1", "Geben Sie hier die erste E-Mail-Adresse ein.");
-$GLOBALS["TL_LANG"]["tl_adressen"]["email2"] = array("E-Mail 2", "Geben Sie hier die zweite E-Mail-Adresse ein.");
-$GLOBALS["TL_LANG"]["tl_adressen"]["email3"] = array("E-Mail 3", "Geben Sie hier die dritte E-Mail-Adresse ein.");
-$GLOBALS["TL_LANG"]["tl_adressen"]["email4"] = array("E-Mail 4", "Geben Sie hier die vierte E-Mail-Adresse ein.");
-$GLOBALS["TL_LANG"]["tl_adressen"]["email5"] = array("E-Mail 5", "Geben Sie hier die fünfte E-Mail-Adresse ein.");
-$GLOBALS["TL_LANG"]["tl_adressen"]["email6"] = array("E-Mail 6", "Geben Sie hier die sechste E-Mail-Adresse ein.");
+$GLOBALS['TL_LANG']['tl_adressen']['email_legende'] = 'E-Mail';
+$GLOBALS['TL_LANG']['tl_adressen']['email_view'] = array('E-Mail veröffentlichen', 'E-Mail-Adressen im Frontend anzeigen');
+$GLOBALS['TL_LANG']['tl_adressen']['email1'] = array('E-Mail 1', 'Geben Sie hier die erste E-Mail-Adresse ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['email2'] = array('E-Mail 2', 'Geben Sie hier die zweite E-Mail-Adresse ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['email3'] = array('E-Mail 3', 'Geben Sie hier die dritte E-Mail-Adresse ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['email4'] = array('E-Mail 4', 'Geben Sie hier die vierte E-Mail-Adresse ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['email5'] = array('E-Mail 5', 'Geben Sie hier die fünfte E-Mail-Adresse ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['email6'] = array('E-Mail 6', 'Geben Sie hier die sechste E-Mail-Adresse ein.');
 
-$GLOBALS["TL_LANG"]["tl_adressen"]["emails_legend"] = "E-Mails (neues Format)";
-$GLOBALS["TL_LANG"]["tl_adressen"]["emails"] = array("Adressen", "");
-$GLOBALS["TL_LANG"]["tl_adressen"]["emails_public_mail"] = array("E-Mail öffentlich", "E-Mail-Adresse im Frontend anzeigen");
-$GLOBALS["TL_LANG"]["tl_adressen"]["emails_mail"] = array("Adresse", "Geben Sie hier die E-Mail-Adresse ein.");
+$GLOBALS['TL_LANG']['tl_adressen']['emails_legend'] = 'E-Mails (neues Format)';
+$GLOBALS['TL_LANG']['tl_adressen']['emails'] = array('Adressen', '');
+$GLOBALS['TL_LANG']['tl_adressen']['emails_public_mail'] = array('E-Mail öffentlich', 'E-Mail-Adresse im Frontend anzeigen');
+$GLOBALS['TL_LANG']['tl_adressen']['emails_mail'] = array('Adresse', 'Geben Sie hier die E-Mail-Adresse ein.');
 
-$GLOBALS["TL_LANG"]["tl_adressen"]["bank_legend"] = "Bankverbindung";
+$GLOBALS['TL_LANG']['tl_adressen']['bank_legend'] = 'Bankverbindung';
+$GLOBALS['TL_LANG']['tl_adressen']['inhaber'] = array('Kontoinhaber', 'Name des Kontoinhabers');
+$GLOBALS['TL_LANG']['tl_adressen']['iban'] = array('IBAN', 'IBAN');
+$GLOBALS['TL_LANG']['tl_adressen']['bic'] = array('BIC', 'BIC');
 
-$GLOBALS["TL_LANG"]["tl_adressen"]["inhaber"][0] = "Kontoinhaber";
-$GLOBALS["TL_LANG"]["tl_adressen"]["inhaber"][1] = "Name des Kontoinhabers";
+$GLOBALS['TL_LANG']['tl_adressen']['web_legende'] = 'Homepage und soziale Netzwerke';
+$GLOBALS['TL_LANG']['tl_adressen']['homepage'] = array('Homepage', 'Geben Sie hier die URL der Internet-Homepage ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['facebook'] = array('Facebook', 'Geben Sie hier den Facebook-Nutzernamen ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['twitter'] = array('Twitter', 'Geben Sie hier den Twitter-Benutzernamen ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['instagram'] = array('Instagram', 'Geben Sie hier den Instagram-Benutzernamen ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['skype'] = array('Skype', 'Geben Sie hier den Skype-Benutzernamen ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['whatsapp'] = array('WhatsApp', 'Geben Sie hier die WhatsApp-Telefonnummer ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['threema'] = array('Threema', 'Geben Sie hier die Threema-ID ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['telegram'] = array('Telegram', 'Geben Sie hier die Telegram-Benutzernamen ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['irc'] = array('IRC', 'Geben Sie hier die IRC-ID ein.');
 
-$GLOBALS["TL_LANG"]["tl_adressen"]["iban"][0] = "IBAN";
-$GLOBALS["TL_LANG"]["tl_adressen"]["iban"][1] = "IBAN";
+$GLOBALS['TL_LANG']['tl_adressen']['image_legend'] = 'Foto';
+$GLOBALS['TL_LANG']['tl_adressen']['singleSRC'] = array('Datei', 'Bitte wählen Sie eine Datei aus dem Dateisystem.');
 
-$GLOBALS["TL_LANG"]["tl_adressen"]["bic"][0] = "BIC";
-$GLOBALS["TL_LANG"]["tl_adressen"]["bic"][1] = "BIC";
+$GLOBALS['TL_LANG']['tl_adressen']['text_legende'] = 'Visitenkarte';
+$GLOBALS['TL_LANG']['tl_adressen']['text'] = array('Visitenkarte', 'Geben Sie hier weitere öffentliche Informationen zur Adresse ein.');
 
-$GLOBALS["TL_LANG"]["tl_adressen"]["web_legende"] = "Soziale Netzwerke und Homepage";
+$GLOBALS['TL_LANG']['tl_adressen']['info_legende'] = 'Bemerkungen/Hinweise';
+$GLOBALS['TL_LANG']['tl_adressen']['info'] = array('Bemerkungen', 'Geben Sie hier interne Bemerkungen zur Adresse ein.');
+$GLOBALS['TL_LANG']['tl_adressen']['links'] = array('Verlinkungen', 'Die Adresse ist in den hier genannten Seiten eingetragen. Diese Übersicht wird einmal nachts aktualisiert.');
+$GLOBALS['TL_LANG']['tl_adressen']['source'] = array('Quelle', 'Ursprüngliche Quelle der Daten');
 
-$GLOBALS["TL_LANG"]["tl_adressen"]["homepage"][0] = "Homepage";
-$GLOBALS["TL_LANG"]["tl_adressen"]["homepage"][1] = "Geben Sie hier die URL der Internet-Homepage ein.";
+$GLOBALS['TL_LANG']['tl_adressen']['aktiv_legende'] = 'Aktivieren';
+$GLOBALS['TL_LANG']['tl_adressen']['aktiv'] = array('Aktiv', 'Aktivieren oder deaktivieren Sie hier die Adresse.');
 
-$GLOBALS["TL_LANG"]["tl_adressen"]["facebook"][0] = "Facebook";
-$GLOBALS["TL_LANG"]["tl_adressen"]["facebook"][1] = "Geben Sie hier den Facebook-Nutzernamen ein. Das einleitende facebook.com/ wird nicht benötigt.";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["twitter"][0] = "Twitter";
-$GLOBALS["TL_LANG"]["tl_adressen"]["twitter"][1] = "Geben Sie hier die Twitter-ID ein.";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["google"][0] = "Google+";
-$GLOBALS["TL_LANG"]["tl_adressen"]["google"][1] = "Geben Sie hier die Google+-ID ein.";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["icq"][0] = "ICQ";
-$GLOBALS["TL_LANG"]["tl_adressen"]["icq"][1] = "Geben Sie hier die ICQ-ID ein.";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["yahoo"][0] = "Yahoo";
-$GLOBALS["TL_LANG"]["tl_adressen"]["yahoo"][1] = "Geben Sie hier die Yahoo-ID ein.";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["aim"][0] = "AIM";
-$GLOBALS["TL_LANG"]["tl_adressen"]["aim"][1] = "Geben Sie hier die AIM-ID ein.";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["msn"][0] = "MSN";
-$GLOBALS["TL_LANG"]["tl_adressen"]["msn"][1] = "Geben Sie hier die MSN-ID ein.";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["irc"][0] = "IRC";
-$GLOBALS["TL_LANG"]["tl_adressen"]["irc"][1] = "Geben Sie hier die IRC-ID ein.";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["image_legend"] = "Bild";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["addImage"][0] = "Ein Bild hinzufügen";
-$GLOBALS["TL_LANG"]["tl_adressen"]["addImage"][1] = "Der Adresse ein Bild hinzufügen.";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["singleSRC"][0] = "Quelldatei";
-$GLOBALS["TL_LANG"]["tl_adressen"]["singleSRC"][1] = "Bitte wählen Sie eine Datei oder einen Ordner aus der Dateiübersicht.";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["alt"][0] = "Alternativer Text";
-$GLOBALS["TL_LANG"]["tl_adressen"]["alt"][1] = "Hier können Sie einen alternativen Text für das Bild eingeben (&lt;em&gt;alt&lt;/em&gt;-Attribut).";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["size"][0] = "Bildbreite und Bildhöhe";
-$GLOBALS["TL_LANG"]["tl_adressen"]["size"][1] = "Hier können Sie die Abmessungen des Bildes und den Skalierungsmodus festlegen.";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["imagemargin"][0] = "Bildabstand";
-$GLOBALS["TL_LANG"]["tl_adressen"]["imagemargin"][1] = "Hier können Sie den oberen, rechten, unteren und linken Bildabstand eingeben.";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["imageUrl"][0] = "Bildlink-Adresse";
-$GLOBALS["TL_LANG"]["tl_adressen"]["imageUrl"][1] = "Eine eigene Bildlink-Adresse überschreibt den Lightbox-Link, so dass das Bild nicht mehr in der Großansicht dargestellt werden kann.";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["fullsize"][0] = "Großansicht/Neues Fenster";
-$GLOBALS["TL_LANG"]["tl_adressen"]["fullsize"][1] = "Großansicht des Bildes in einer Lightbox bzw. den Link in einem neuem Browserfenster öffnen.";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["caption"][0] = "Bildunterschrift";
-$GLOBALS["TL_LANG"]["tl_adressen"]["caption"][1] = "Hier können Sie einen kurzen Text eingeben, der unterhalb des Bildes angezeigt wird.";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["floating"][0] = "Bildausrichtung";
-$GLOBALS["TL_LANG"]["tl_adressen"]["floating"][1] = "Bitte legen Sie fest, wie das Bild ausgerichtet werden soll.";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["text_legende"] = "Visitenkarte";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["text"][0] = "Visitenkarte";
-$GLOBALS["TL_LANG"]["tl_adressen"]["text"][1] = "Geben Sie hier weitere öffentliche Informationen zur Adresse ein.";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["info_legende"] = "Bemerkungen/Hinweise";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["info"][0] = "Bemerkungen";
-$GLOBALS["TL_LANG"]["tl_adressen"]["info"][1] = "Geben Sie hier interne Bemerkungen zur Adresse ein.";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["links"][0] = "Verlinkungen";
-$GLOBALS["TL_LANG"]["tl_adressen"]["links"][1] = "Die Adresse ist in den hier genannten Seiten eingetragen. Diese Übersicht wird einmal nachts aktualisiert.";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["source"][0] = "Quelle";
-$GLOBALS["TL_LANG"]["tl_adressen"]["source"][1] = "Ursprüngliche Quelle der Daten";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["aktiv_legende"] = "Aktivieren";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["aktiv"][0] = "Aktiv";
-$GLOBALS["TL_LANG"]["tl_adressen"]["aktiv"][1] = "Aktivieren oder deaktivieren Sie hier die Adresse.";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["alias_legende"] = "Alias";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["alias"][0] = "Alias";
-$GLOBALS["TL_LANG"]["tl_adressen"]["alias"][1] = "Das Alias wird automatisch beim Speichern aus Vor- und Nachname generiert, wenn das Feld leer ist.";
-
-$GLOBALS["TL_LANG"]["tl_adressen"]["searchstring"][0] = "Volltextsuche";
-$GLOBALS["TL_LANG"]["tl_adressen"]["searchstring"][1] = "Daten für Volltextsuche";
+$GLOBALS['TL_LANG']['tl_adressen']['searchstring'] = array('Volltextsuche', 'Daten für Volltextsuche');
 
 
 /**
@@ -168,28 +102,12 @@ $GLOBALS["TL_LANG"]["tl_adressen"]["searchstring"][1] = "Daten für Volltextsuch
  */
 
 $GLOBALS['TL_LANG']['tl_adressen']['categories'] = array('Kategorien', 'Kategorien für die Adressen bearbeiten');
-
-$GLOBALS['TL_LANG']['tl_adressen']['import'][0] = 'CSV-Import';
-$GLOBALS['TL_LANG']['tl_adressen']['import'][1] = 'Daten aus einer CSV-Datei importieren';
-
-$GLOBALS['TL_LANG']['tl_adressen']['export'][0] = 'CSV-Export';
-$GLOBALS['TL_LANG']['tl_adressen']['export'][1] = 'Daten als CSV-Datei exportieren';
-
-$GLOBALS['TL_LANG']['tl_adressen']['new'][0] = "Neue Adresse";
-$GLOBALS['TL_LANG']['tl_adressen']['new'][1] = "Neue Adresse anlegen";
-
-$GLOBALS['TL_LANG']['tl_adressen']['edit'][0] = "Adresse bearbeiten";
-$GLOBALS['TL_LANG']['tl_adressen']['edit'][1] = "Adresse %s bearbeiten";
-
-$GLOBALS['TL_LANG']['tl_adressen']['copy'][0] = "Adresse kopieren";
-$GLOBALS['TL_LANG']['tl_adressen']['copy'][1] = "Adresse %s kopieren";
-
-$GLOBALS['TL_LANG']['tl_adressen']['delete'][0] = "Adresse löschen";
-$GLOBALS['TL_LANG']['tl_adressen']['delete'][1] = "Adresse %s löschen";
-
-$GLOBALS['TL_LANG']['tl_adressen']['toggle'][0] = "Adresse aktivieren/deaktivieren";
-$GLOBALS['TL_LANG']['tl_adressen']['toggle'][1] = "Adresse %s aktivieren/deaktivieren";
-
-$GLOBALS['TL_LANG']['tl_adressen']['show'][0] = "Adressdetails anzeigen";
-$GLOBALS['TL_LANG']['tl_adressen']['show'][1] = "Details der Adresse %s anzeigen";
+$GLOBALS['TL_LANG']['tl_adressen']['import'] = array('CSV-Import', 'Daten aus einer CSV-Datei importieren');
+$GLOBALS['TL_LANG']['tl_adressen']['export'] = array('CSV-Export', 'Daten als CSV-Datei exportieren');
+$GLOBALS['TL_LANG']['tl_adressen']['new'] = array('Neue Adresse', 'Neue Adresse anlegen');
+$GLOBALS['TL_LANG']['tl_adressen']['edit'] = array('Adresse bearbeiten', 'Adresse %s bearbeiten');
+$GLOBALS['TL_LANG']['tl_adressen']['copy'] = array('Adresse kopieren', 'Adresse %s kopieren');
+$GLOBALS['TL_LANG']['tl_adressen']['delete'] = array('Adresse löschen', 'Adresse %s löschen');
+$GLOBALS['TL_LANG']['tl_adressen']['toggle'] = array('Adresse aktivieren/deaktivieren', 'Adresse %s aktivieren/deaktivieren');
+$GLOBALS['TL_LANG']['tl_adressen']['show'] = array('Adressdetails anzeigen', 'Details der Adresse %s anzeigen');