+++ src/Resources/contao/languages/de/tl_content.php
@@ -1,29 +1,21 @@
 <?php 
 
-$GLOBALS['TL_LANG']['tl_content']['adresse_id']['0'] = "Adresse auswählen";
-$GLOBALS['TL_LANG']['tl_content']['adresse_id']['1'] = "Bitte wählen Sie die anzuzeigende Adresse aus.";
-$GLOBALS['TL_LANG']['tl_content']['adresse_tpl']['0'] = "Template";
-$GLOBALS['TL_LANG']['tl_content']['adresse_tpl']['1'] = "Bitte wählen Sie das Frontend-Template aus.";
-$GLOBALS['TL_LANG']['tl_content']['adresse_funktion']['0'] = "Funktion/Amt";
-$GLOBALS['TL_LANG']['tl_content']['adresse_funktion']['1'] = "Erscheint vor dem Namen in Fettschrift.";
-$GLOBALS['TL_LANG']['tl_content']['adresse_zusatz']['0'] = "Zusatztext";
-$GLOBALS['TL_LANG']['tl_content']['adresse_zusatz']['1'] = "Erscheint zwischen Funktion/Amt und Name. HTML ist erlaubt.";
-$GLOBALS['TL_LANG']['tl_content']['adresse_bildvorschau']['0'] = "Standardbild";
-$GLOBALS['TL_LANG']['tl_content']['adresse_bildvorschau']['1'] = "Das Bild wird im Original, aber ggfs. skaliert angezeigt.";
-$GLOBALS['TL_LANG']['tl_content']['adresse_bildvorschau_fehlt']['0'] = "Kein Standardbild vorhanden";
-$GLOBALS['TL_LANG']['tl_content']['adresse_bildvorschau_fehlt']['1'] = "";
-$GLOBALS['TL_LANG']['tl_content']['adresse_viewfoto']['0'] = "Bild anzeigen";
-$GLOBALS['TL_LANG']['tl_content']['adresse_viewfoto']['1'] = "Standardbild bzw. alternatives Bild anzeigen.";
-$GLOBALS['TL_LANG']['tl_content']['adresse_selectmails']['0'] = "Nur bestimmte E-Mail-Adressen anzeigen";
-$GLOBALS['TL_LANG']['tl_content']['adresse_selectmails']['1'] = "Anzeige der E-Mail-Adressen einschränken";
-$GLOBALS['TL_LANG']['tl_content']['adresse_mails']['0'] = "E-Mail-Adressen auswählen";
-$GLOBALS['TL_LANG']['tl_content']['adresse_mails']['1'] = "Nur aktivierte E-Mail-Adressen werden angezeigt.";
-
 $GLOBALS['TL_LANG']['tl_content']['adresse_legend'] = 'Einstellungen';
-$GLOBALS['TL_LANG']['tl_content']['adressbild_legend'] = 'Alternatives Bild (überschreibt Standardbild)';
+$GLOBALS['TL_LANG']['tl_content']['adresse_id'] = array('Adresse auswählen', 'Bitte wählen Sie die anzuzeigende Adresse aus.');
+$GLOBALS['TL_LANG']['tl_content']['adresse_funktion'] = array('Funktion/Amt', 'Erscheint vor dem Namen in Fettschrift.');
+$GLOBALS['TL_LANG']['tl_content']['adresse_zusatz'] = array('Zusatztext', 'Erscheint zwischen Funktion/Amt und Name. HTML ist erlaubt.');
+$GLOBALS['TL_LANG']['tl_content']['adresse_selectmails'] = array('Nur bestimmte E-Mail-Adressen anzeigen', 'Anzeige der E-Mail-Adressen einschränken');
+$GLOBALS['TL_LANG']['tl_content']['adresse_mails'] = array('E-Mail-Adressen auswählen', 'Nur aktivierte E-Mail-Adressen werden angezeigt.');
+
+$GLOBALS['TL_LANG']['tl_content']['adressefoto_legend'] = 'Foto';
+$GLOBALS['TL_LANG']['tl_content']['adresse_viewFoto'] = array('Foto anzeigen', 'Foto im Frontend anzeigen');
+$GLOBALS['TL_LANG']['tl_content']['adresse_bildvorschau'] = array('Zugeordnetes Foto', 'Das Bild wird in der Frontendgröße angezeigt.');
+$GLOBALS['TL_LANG']['tl_content']['adresse_singleSRC'] = array('Foto ersetzen durch:', 'Datei aus dem Dateisystem wählen');
+
+$GLOBALS['TL_LANG']['tl_content']['adresstemplate_legend'] = 'Alternatives Template';
+$GLOBALS['TL_LANG']['tl_content']['adresse_alttemplate'] = array('Aktivieren', 'Alternatives Template aktivieren');
+$GLOBALS['TL_LANG']['tl_content']['adresse_tpl'] = array('Template', 'Bitte wählen Sie das zu verwendende Frontend-Template aus.');
 
 // Sonstige 
-$GLOBALS['TL_LANG']['tl_content']['adresse_bildvorschau_leer'] = "Fehler: Pfad nicht gefunden!";
+$GLOBALS['TL_LANG']['tl_content']['adresse_bildvorschau_leer'] = 'Fehler: Pfad nicht gefunden!';
 $GLOBALS['TL_LANG']['tl_content']['adresse_nichtaktiv'] = ' (deaktiviert)';
-
-?>